+++ lib/arel/collectors/elasticsearch_query.rb
@@ -5,21 +5,22 @@ require 'elasticsearch_record/query'
 module Arel # :nodoc: all
   module Collectors
     class ElasticsearchQuery < ::ElasticsearchRecord::Query
+
+      # required for ActiveRecord
+      attr_accessor :preparable
+
       def initialize
-        super
-
         # force initialize a body as hash
-        @body ||= {}
+        super(body: {})
 
-        # stack is used to point on a specific hash node
-        @stack = []
+        # @binds = []
+        @bind_index = 1
       end
 
-      # send a proposal to this query -
+      # send a proposal to this query
       # @param [Symbol] action - the claim action
       # @param [Array] args - args to claim
-      # @param [Proc] block - a optional block to create a nested scope on the current provided key
-      def claim(action, *args, &block)
+      def claim(action, *args)
         case action
         when :index
           # change the index name
@@ -42,104 +43,60 @@ module Arel # :nodoc: all
         when :assign
           # calls a assign on the body
           assign(*args)
-
-          if block_given?
-            raise "Unsupported claim key: '#{args[0]}' for provided block. Provide a Symbol as key!" unless args[0].is_a?(Symbol)
-
-            scope(args[0]) do
-              block.call
-            end
-          end
         else
           raise "Unsupported claim action: #{action}"
         end
       end
 
       def <<(claim)
-        Debugger.debug(claim,"receiving a claim")
-
-        self.claim(claim[0], *claim[1], &claim[2])
+        self.claim(claim[0], *claim[1])
       end
 
-      def current
-        @current ||= @stack.any? ? @body.dig(*@stack) : @body
-      end
-
-      def current=(value)
-        if (key = @stack.last).present?
-          __send__(:parent)[key] = value
-        else
-          @body = value
-        end
-
-        # IMPORTANT! Reset current
-        @current = nil
-      end
-
+      # used by the +Arel::Visitors::Elasticsearch#compile+ method (and default Arel visitors)
+      # todo: maybe return arguments with :_meta information instead of self ...
       def value
+        self
+      end
+
+      # IMPORTANT: For SQL defaults (see @ Arel::Collectors::SubstituteBinds) a value
+      # will +not+ be directly assigned (see @ Arel::Visitors::ToSql#visit_Arel_Nodes_HomogeneousIn).
+      # instead it will be send as bind and then re-delegated to the SQL collector.
+      #
+      # This only works for linear SQL-queries and not nested Hashes
+      # (otherwise we have to collect those binds, and replace them afterwards).
+      #
+      # This will be ignored by the ElasticsearchQuery collector, but supports statement caches on the other side
+      # (see @ ActiveRecord::StatementCache::PartialQueryCollector)
+      def add_bind(bind, &block)
+        @bind_index += 1
+
+        self
+      end
+
+      # @see Arel::Collectors::ElasticsearchQuery#add_bind
+      def add_binds(binds, proc_for_binds = nil, &block)
+        @bind_index += binds.size
         self
       end
 
       private
 
-      def parent
-        return @body if @stack.length <= 1
-        @body.dig(*@stack[0..-2])
-      end
-
-      def assign(*args)
+      # calls a assign on the body
+      def assign(key, value)
         # check for special provided key, to claim through an assign
-        if args[0] == :__query__
-          if args[1].is_a?(Array)
-            args[1].each do |arg|
-              key = arg.keys.first
-              claim(key, arg[key])
+        if key == :__claim__
+          if value.is_a?(Array)
+            value.each do |arg|
+              vkey = arg.keys.first
+              claim(vkey, arg[vkey])
             end
           else
-            key = args[1].keys.first
-            claim(key, args[1][key])
-          end
-
-          return self
-        end
-
-        if self.current.is_a?(Array)
-          if args[0].is_a?(Array)
-            self.current += args[0]
-          else
-            self.current << args[0]
-          end
-        elsif self.current.is_a?(Hash)
-          if args[0].is_a?(Hash)
-            self.current = self.current.merge(args[0])
-          elsif self.current[args[0]].is_a?(Hash) && args[1].is_a?(Hash)
-            self.current[args[0]] = self.current[args[0]].merge(args[1])
-          elsif args[1].nil?
-            self.current.delete(args[0])
-          else
-            self.current[args[0]] = args[1]
-          end
-        elsif self.current.is_a?(String)
-          if args[0].is_a?(Array)
-            self.current = self.current.to_s + args[0].map(&:to_s).join
-          else
-            self.current = self.current.to_s + args[0].to_s
+            vkey = value.keys.first
+            claim(vkey, value[vkey])
           end
         else
-          self.current = args[0]
+          @body[key] = value
         end
-
-        self
-      end
-
-      def scope(key)
-        @stack << key
-        @current = nil
-
-        yield
-
-        @stack.pop
-        @current = nil
       end
     end
   end