+++ lib/elasticsearch_record/gem_version.rb
@@ -7,9 +7,9 @@ module ElasticsearchRecord
   end
 
   module VERSION
-    MAJOR = 0
-    MINOR = 1
-    TINY  = 2
+    MAJOR = 1
+    MINOR = 0
+    TINY  = 0
     PRE   = nil
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")