+++ lib/elasticsearch_record/patches/arel/update_statement_patch.rb
@@ -7,6 +7,25 @@ module ElasticsearchRecord
   module Patches
     module Arel
       module UpdateStatementPatch
+        def self.included(base)
+          base.send(:prepend, PrependMethods)
+        end
+
+        module PrependMethods
+          def initialize_copy(other)
+            super
+            @configure  = @configure&.deep_dup
+            @queries  = @queries&.clone
+          end
+
+          def hash
+            [super, @configure, @queries].hash
+          end
+
+          def eql?(other)
+            super && kind == other.kind && configure == other.configure && queries == other.queries
+          end
+        end
 
         def kind
           @kind
@@ -30,20 +49,6 @@ module ElasticsearchRecord
 
         def configure=(value)
           @configure = value
-        end
-
-        def initialize_copy(other)
-          super
-          @configure  = @configure&.deep_dup
-          @queries  = @queries&.clone
-        end
-
-        def hash
-          [super, @configure, @queries].hash
-        end
-
-        def eql?(other)
-          super && kind == other.kind && configure == other.configure && queries == other.queries
         end
       end
     end