+++ lib/elasticsearch_record/relation/result_methods.rb
@@ -96,7 +96,7 @@ module ElasticsearchRecord
         relation.offset!(nil).limit!(nil)
 
         # remove the 'index' from the query arguments (pit doesn't like that)
-        relation.configure!(:__query__, { index: nil })
+        relation.configure!(:__claim__, { index: nil })
 
         # prepare a total results array
         results       = []